+++ src/elements/FieldsetRow.tsx
@@ -5,7 +5,8 @@ import PropTypes from "prop-types";
 
 export type FieldsetRowProps = {
   children: any,
-  alignRight?: boolean
+  alignRight?: boolean,
+  testId?: string,
 };
 
 export const FieldsetRowPropTypes = {
@@ -13,11 +14,11 @@ export const FieldsetRowPropTypes = {
   alignRight: PropTypes.bool
 }
 
-function FieldsetRow({children, alignRight=false}: FieldsetRowProps) {
+function FieldsetRow({ children, alignRight = false, testId }: FieldsetRowProps) {
   if (alignRight) {
-    return <div className="mt-5 text-right md:space-x-3 md:block flex flex-col-reverse">{children}</div>
+    return <div data-testid={testId} className="mt-5 text-right md:space-x-3 md:block flex flex-col-reverse">{children}</div>
   }
-  return <div className="md:flex flex-row md:space-x-4 w-full text-xs">{children}</div>;
+  return <div data-testid={testId} className="md:flex flex-row md:space-x-4 w-full text-xs">{children}</div>;
 }
 
 FieldsetRow.propTypes = FieldsetRowPropTypes;